pyproject.toml
[tool.poetry]
name = "chatbot-back"
version = "0.1.0"
description = ""
authors = ["likeRudin <likeRudin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg = "^3.2.3"
psycopg2-binary = "^2.9.10"
django-cors-headers = "^4.5.0"
pyjwt = "^2.9.0"
openai = "^1.52.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
